requirements.txt
# Python dependencies for Daily Briefing Bot

# Core dependencies
aiohttp>=3.8.0
schedule>=1.2.0
python-dotenv>=1.0.0

# Weather API dependencies
openmeteo-requests>=1.0.0
requests-cache>=1.0.0
retry-requests>=2.0.0
numpy>=1.20.0
pandas>=1.3.0

# Optional: For real calendar integration
# google-auth>=2.0.0
# google-auth-oauthlib>=1.0.0
# google-auth-httplib2>=0.1.0
# google-api-python-client>=2.0.0
# msal>=1.20.0

# Development dependencies
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
